# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Forensic Document Examiner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DocumentUpload {
  /** The PDF document to analyze */
  file: Blob;
  /** What the submitting party claims this document is (e.g. "a letter from First National Bank dated March 2022") */
  claimedIdentity: string;
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  info: 'info',
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Finding {
  /** Category of the finding (e.g. metadata, software, timestamp, content, fonts) */
  category: string;
  severity: FindingSeverity;
  title: string;
  description: string;
}

export interface DocumentMetadata {
  /** @nullable */
  author?: string | null;
  /** @nullable */
  creator?: string | null;
  /** @nullable */
  producer?: string | null;
  /** @nullable */
  creationDate?: string | null;
  /** @nullable */
  modificationDate?: string | null;
  /** @nullable */
  pageCount?: number | null;
  fileSize?: number;
  encrypted?: boolean;
  hasDigitalSignature?: boolean;
  /** @nullable */
  pdfVersion?: string | null;
}

export type DocumentAnalysisStatus = typeof DocumentAnalysisStatus[keyof typeof DocumentAnalysisStatus];


export const DocumentAnalysisStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  complete: 'complete',
  error: 'error',
} as const;

/**
 * @nullable
 */
export type DocumentAnalysisVerdict = typeof DocumentAnalysisVerdict[keyof typeof DocumentAnalysisVerdict] | null;


export const DocumentAnalysisVerdict = {
  authentic: 'authentic',
  suspicious: 'suspicious',
  likely_forged: 'likely_forged',
  inconclusive: 'inconclusive',
} as const;

export interface DocumentAnalysis {
  id: number;
  status: DocumentAnalysisStatus;
  fileName: string;
  fileSize: number;
  claimedIdentity: string;
  /** @nullable */
  verdict?: DocumentAnalysisVerdict;
  /**
     * 0-100 confidence in the verdict
     * @nullable
     */
  confidenceScore?: number | null;
  /** @nullable */
  summary?: string | null;
  findings?: Finding[];
  metadata?: DocumentMetadata | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type DocumentStatsByVerdict = {
  authentic?: number;
  suspicious?: number;
  likely_forged?: number;
  inconclusive?: number;
};

export type DocumentStatsByStatus = {
  pending?: number;
  analyzing?: number;
  complete?: number;
  error?: number;
};

export interface DocumentStats {
  total: number;
  byVerdict: DocumentStatsByVerdict;
  byStatus: DocumentStatsByStatus;
  /** Documents analyzed in the last 7 days */
  recentCount: number;
}