# lib/api-zod/src/generated/types/documentAnalysis.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Forensic Document Examiner API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentAnalysisStatus } from './documentAnalysisStatus';
import type { DocumentAnalysisVerdict } from './documentAnalysisVerdict';
import type { DocumentMetadata } from './documentMetadata';
import type { Finding } from './finding';

export interface DocumentAnalysis {
  id: number;
  status: DocumentAnalysisStatus;
  fileName: string;
  fileSize: number;
  claimedIdentity: string;
  /** @nullable */
  verdict?: DocumentAnalysisVerdict;
  /**
     * 0-100 confidence in the verdict
     * @nullable
     */
  confidenceScore?: number | null;
  /** @nullable */
  summary?: string | null;
  findings?: Finding[];
  metadata?: DocumentMetadata | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}