# lib/api-zod/src/generated/types/documentMetadata.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Forensic Document Examiner API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentMetadata {
  /** @nullable */
  author?: string | null;
  /** @nullable */
  creator?: string | null;
  /** @nullable */
  producer?: string | null;
  /** @nullable */
  creationDate?: string | null;
  /** @nullable */
  modificationDate?: string | null;
  /** @nullable */
  pageCount?: number | null;
  fileSize?: number;
  encrypted?: boolean;
  hasDigitalSignature?: boolean;
  /** @nullable */
  pdfVersion?: string | null;
}