# lib/api-zod/src/generated/types/documentStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Forensic Document Examiner API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatsByStatus } from './documentStatsByStatus';
import type { DocumentStatsByVerdict } from './documentStatsByVerdict';

export interface DocumentStats {
  total: number;
  byVerdict: DocumentStatsByVerdict;
  byStatus: DocumentStatsByStatus;
  /** Documents analyzed in the last 7 days */
  recentCount: number;
}