# lib/api-zod/src/generated/types/documentUpload.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Forensic Document Examiner API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentUpload {
  /** The PDF document to analyze */
  file: Blob;
  /** What the submitting party claims this document is (e.g. "a letter from First National Bank dated March 2022") */
  claimedIdentity: string;
}