# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Forensic Document Examiner API
 * OpenAPI spec version: 0.1.0
 */
import type { FindingSeverity } from './findingSeverity';

export interface Finding {
  /** Category of the finding (e.g. metadata, software, timestamp, content, fonts) */
  category: string;
  severity: FindingSeverity;
  title: string;
  description: string;
}